refactored queries
